+++ src/app/user/page.tsx
@@ -1,7 +1,7 @@
 import Sidebar from "@/common/component/Sidebar";
 import { sidebarWrapper } from "@/styles/shared.css";
-import UserCard from "@/view/user/index/component/UserCard";
-import { userCardWrapper } from "@/view/user/index/component/UserCard/index.css";
+import UserCard from "@/view/user/index/UserCard";
+import { userCardWrapper } from "@/view/user/index/UserCard/index.css";
 
 const UserDashboardPage = () => {
   return (
