+++ src/app/user/setting/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import Sidebar from "@/common/component/Sidebar";
 import { sidebarWrapper } from "@/styles/shared.css";
-import SettingStep from "@/view/user/setting/component/SettingStep";
+import SettingStep from "@/view/user/setting/SettingStep";
 import type { SettingSteps } from "@/view/user/setting/type";
 import { useState } from "react";
