+++ src/view/group/dashboard/GroupSidebar/index.css.ts
