--- src/view/group/dashboard/component/GroupSidebar/CircleNumber.tsx
@@ -1,8 +1,0 @@
-import { circleNumberStyle } from "@/view/group/dashboard/component/GroupSidebar/index.css";
-import type { PropsWithChildren } from "react";
-
-const CircleNumber = ({ children }: PropsWithChildren) => {
-  return <div className={circleNumberStyle}>{children}</div>;
-};
-
-export default CircleNumber;
