+++ src/view/group/index/MemberAvatar/index.css.tsx
