+++ src/view/group/index/ProblemSidebar/index.css.ts
