+++ src/view/onboarding/Footer/index.css.ts
