+++ src/view/onboarding/Footer/index.tsx
