+++ src/view/onboarding/Section/FadeInImage.tsx
