+++ src/view/onboarding/Section/Section1.tsx
