+++ src/view/onboarding/Section/Section2.tsx
