+++ src/view/onboarding/Section/Section3.tsx
