+++ src/view/onboarding/Section/Text.tsx
