+++ src/view/onboarding/Section/index.css.ts
