+++ src/view/onboarding/Section/index.tsx
