+++ src/view/user/join-group/GroupInfoCard/GroupDateInfo.css.ts
