+++ src/view/user/join-group/GroupInfoCard/index.css.ts
